Internship_Final/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.3.2
TTS==0.22.0
sentence-transformers==2.2.2
chromadb==0.4.18
scikit-learn==1.3.2
numpy==1.24.3